@@ -2,12 +2,13 @@ name: connectivity_plus_example
 description: Demonstrates how to use the connectivity_plus plugin.
 
 environment:
-  sdk: '>=2.12.0 <4.0.0'
+  sdk: '>=3.3.1 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  connectivity_plus: ^5.0.2
+  connectivity_plus:
+      path: ../ 
 
 dev_dependencies:
   flutter_driver:
