@@ -42,9 +42,9 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  ConnectivityResult _connectionStatus = ConnectivityResult.none;
+  List<ConnectivityResult> _connectionStatus = [ConnectivityResult.none];
   final Connectivity _connectivity = Connectivity();
-  late StreamSubscription<ConnectivityResult> _connectivitySubscription;
+  late StreamSubscription<List<ConnectivityResult>> _connectivitySubscription;
 
   @override
   void initState() {
@@ -63,7 +63,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initConnectivity() async {
-    late ConnectivityResult result;
+    late List<ConnectivityResult> result;
     // Platform messages may fail, so we use a try/catch PlatformException.
     try {
       result = await _connectivity.checkConnectivity();
@@ -82,7 +82,7 @@ class _MyHomePageState extends State<MyHomePage> {
     return _updateConnectionStatus(result);
   }
 
-  Future<void> _updateConnectionStatus(ConnectivityResult result) async {
+  Future<void> _updateConnectionStatus(List<ConnectivityResult> result) async {
     setState(() {
       _connectionStatus = result;
     });
@@ -92,11 +92,32 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Connectivity example app'),
+        title: const Text('Connectivity Plus Example'),
         elevation: 4,
       ),
-      body: Center(
-          child: Text('Connection Status: ${_connectionStatus.toString()}')),
+      body: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const Spacer(flex: 2),
+          Text(
+            'Active connection types:',
+            style: Theme.of(context).textTheme.headlineMedium,
+          ),
+          const Spacer(),
+          ListView(
+            shrinkWrap: true,
+            children: List.generate(
+                _connectionStatus.length,
+                (index) => Center(
+                      child: Text(
+                        _connectionStatus[index].toString(),
+                        style: Theme.of(context).textTheme.headlineSmall,
+                      ),
+                    )),
+          ),
+          const Spacer(flex: 2),
+        ],
+      ),
     );
   }
 }
