@@ -2,10 +2,14 @@
 // Use of this source is governed by a BSD-style license that can
 // be found in the LICENSE file.
 
+#if os(iOS)
+import Flutter
+#elseif os(macOS)
 import Cocoa
 import FlutterMacOS
+#endif
 
-public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
+public class SwiftConnectivityPlusPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
   private let connectivityProvider: ConnectivityProvider
   private var eventSink: FlutterEventSink?
 
@@ -16,22 +20,28 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
   }
 
   public static func register(with registrar: FlutterPluginRegistrar) {
+    #if os(iOS)
+    let binaryMessenger = registrar.messenger()
+    #elseif os(macOS)
+    let binaryMessenger = registrar.messenger
+    #endif
+
     let channel = FlutterMethodChannel(
       name: "dev.fluttercommunity.plus/connectivity",
-      binaryMessenger: registrar.messenger)
+      binaryMessenger: binaryMessenger)
 
     let streamChannel = FlutterEventChannel(
       name: "dev.fluttercommunity.plus/connectivity_status",
-      binaryMessenger: registrar.messenger)
+      binaryMessenger: binaryMessenger)
 
     let connectivityProvider: ConnectivityProvider
-    if #available(macOS 10.14, *) {
+    if #available(iOS 12, macOS 10.14, *) {
       connectivityProvider = PathMonitorConnectivityProvider()
     } else {
       connectivityProvider = ReachabilityConnectivityProvider()
     }
 
-    let instance = ConnectivityPlugin(connectivityProvider: connectivityProvider)
+    let instance = SwiftConnectivityPlusPlugin(connectivityProvider: connectivityProvider)
     streamChannel.setStreamHandler(instance)
 
     registrar.addMethodCallDelegate(instance, channel: channel)
@@ -45,7 +55,7 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
     case "check":
-      result(statusFrom(connectivityType: connectivityProvider.currentConnectivityType))
+      result(statusFrom(connectivityTypes: connectivityProvider.currentConnectivityTypes))
     default:
       result(FlutterMethodNotImplemented)
     }
@@ -60,9 +70,15 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
     case .wiredEthernet:
       return "ethernet"
     case .other:
-      return "other"
+        return "other"
     case .none:
       return "none"
+    }
+  }
+  
+  private func statusFrom(connectivityTypes: [ConnectivityType]) -> [String] {
+    return connectivityTypes.map {
+      self.statusFrom(connectivityType: $0)
     }
   }
 
@@ -72,13 +88,14 @@ public class ConnectivityPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
   ) -> FlutterError? {
     eventSink = events
     connectivityProvider.start()
-    connectivityUpdateHandler(connectivityType: connectivityProvider.currentConnectivityType)
+    // Update this to handle a list
+    connectivityUpdateHandler(connectivityTypes: connectivityProvider.currentConnectivityTypes)
     return nil
   }
 
-  private func connectivityUpdateHandler(connectivityType: ConnectivityType) {
+  private func connectivityUpdateHandler(connectivityTypes: [ConnectivityType]) {
     DispatchQueue.main.async {
-      self.eventSink?(self.statusFrom(connectivityType: connectivityType))
+      self.eventSink?(self.statusFrom(connectivityTypes: connectivityTypes))
     }
   }
 
