@@ -19,7 +19,7 @@ class ConnectivityPlusLinuxPlugin extends ConnectivityPlatform {
 
   /// Checks the connection status of the device.
   @override
-  Future<ConnectivityResult> checkConnectivity() async {
+  Future<List<ConnectivityResult>> checkConnectivity() async {
     final client = createClient();
     await client.connect();
     final connectivity = _getConnectivity(client);
@@ -28,38 +28,44 @@ class ConnectivityPlusLinuxPlugin extends ConnectivityPlatform {
   }
 
   NetworkManagerClient? _client;
-  StreamController<ConnectivityResult>? _controller;
+  StreamController<List<ConnectivityResult>>? _controller;
 
   /// Returns a Stream of ConnectivityResults changes.
   @override
-  Stream<ConnectivityResult> get onConnectivityChanged {
-    _controller ??= StreamController<ConnectivityResult>.broadcast(
+  Stream<List<ConnectivityResult>> get onConnectivityChanged {
+    _controller ??= StreamController<List<ConnectivityResult>>.broadcast(
       onListen: _startListenConnectivity,
       onCancel: _stopListenConnectivity,
     );
     return _controller!.stream;
   }
 
-  ConnectivityResult _getConnectivity(NetworkManagerClient client) {
+  List<ConnectivityResult> _getConnectivity(NetworkManagerClient client) {
+    final List<ConnectivityResult> results = [];
     if (client.connectivity == NetworkManagerConnectivityState.none) {
-      return ConnectivityResult.none;
+      results.add(ConnectivityResult.none);
+    } else {
+      if (client.primaryConnectionType.contains('wireless')) {
+        results.add(ConnectivityResult.wifi);
+      }
+      if (client.primaryConnectionType.contains('ethernet')) {
+        results.add(ConnectivityResult.ethernet);
+      }
+      if (client.primaryConnectionType.contains('vpn')) {
+        results.add(ConnectivityResult.vpn);
+      }
+      if (client.primaryConnectionType.contains('bluetooth')) {
+        results.add(ConnectivityResult.bluetooth);
+      }
+      if (client.primaryConnectionType.contains('mobile')) {
+        results.add(ConnectivityResult.mobile);
+      }
+      // Assuming 'other' is a catch-all for unspecified types
+      if (results.isEmpty) {
+        results.add(ConnectivityResult.other);
+      }
     }
-    if (client.primaryConnectionType.contains('wireless')) {
-      return ConnectivityResult.wifi;
-    }
-    if (client.primaryConnectionType.contains('ethernet')) {
-      return ConnectivityResult.ethernet;
-    }
-    if (client.primaryConnectionType.contains('vpn')) {
-      return ConnectivityResult.vpn;
-    }
-    if (client.primaryConnectionType.contains('bluetooth')) {
-      return ConnectivityResult.bluetooth;
-    }
-    if (client.primaryConnectionType.contains('mobile')) {
-      return ConnectivityResult.mobile;
-    }
-    return ConnectivityResult.other;
+    return results;
   }
 
   Future<void> _startListenConnectivity() async {
